Create new-version-square-generatation.bas
After the shorter version I created a version of that with a bit more explanation so it can be used to learn from it. So this is the version with more remarks added. 

The original version I created first was more to experiment with the use of the DIM statement and matrix to see how that could be used. However that one was a bit slow.